Fix `who-master` `None` reply handling
The return value of a `who-master` call is a `string option`. This
patch fixes the request function to handle a `None` reply.

If the server returns `None`, the returned string value is set to
`NULL`, whilst this value is passed to `arakoon_utils_make_string`
unconditionally, which requires a non-NULL value (and is tagged as
such).

Bug found thanks to the LLVM/Clang Analyzer.

Change-Id: I7d55ba97017443b7a66ae66a2cdeb8f3a583ab13