Add NULL checks in public API procedures
Depending on the return value of the specific procedure, -EINVAL is
returned when a NULL argument is provided which should be non-NULL, or
NULL is returned and errno is set to EINVAL in case a pointer type is
returned.

Note some procedures might also return NULL on allocation error, so
always check errno to make sure what's going on.

Change-Id: I3ee2c0bb59877b425fb7205bf964e5ef705ccfd1