Implement multi-address connect
This patch changes `_arakoon_networking_connect` to treat the given
`addrinfo` pointer as a list of addresses to connect to, attempt to
connect to all of them, and return the first successful connection.

This can be used for multi-IP setups (but only when using
`arakoon_cluster_add_node` to add a node to a cluster, providing an
`addrinfo` list manually).

A testcase should be added.

Change-Id: I34052fd28162cca48a9737e33f757b46d094c997
Fixes: ARAKOON-293
See: http://jira.incubaid.com/browse/ARAKOON-293