Fix header file
The header-file change introduced in 540e67b93ca contains an error
which is not caught by a C compiler (and isn't invalid, actually) but
does fail when using a C++ compiler.

Since the header is supposed to be C++-compatible, this needs fixing
(it was an error anyway).

Thanks to Thanks to Bastiaan Stougie for reporting this.

Reported-by: Bastiaan Stougie <bastiaan.stougie@amplidata.com>
See: 540e67b93ca518f9faf1b45fbaeec9df28f79afc

Change-Id: I02a15a701165d932e692951d151ef46818791509