Reset errno in poll-handler loop
When read(2) returns a zero-length result, this might indicate no data
could be read from a socket, which is not an error condition (errno
won't be changed). The errno value could still contain the value set
during a prior iteration (e.g. EINTR or EAGAIN), which results in an
infinite loop.

This commit changes the inner loop to reset errno to 0 when entered.

Thanks to Arne Redlich, Immanuel Litzroth and Bart De Vylder for
pinpointing the issue.

Change-Id: Ib4a2f63697a13d06e3af7e926dc44c69f3db8cb3
Fixes: ARAKOON-324
See: http://jira.incubaid.com/browse/ARAKOON-324