Fix connect with infinite timeout
If no timeout / an infinite timeout is set, _arakoon_networking_connect()
sets the poll() timeout to 0. If the initial connect() call fails with
EINPROGRESS the subsequent poll() will hence immediately return with a good
chance of the socket not being writable yet, failing the whole connect
operation.

Fix this by setting the timeout to -1 in that case, indicating that we
want to wait as long as it takes for the socket to become connected.

Change-Id: Ia89187e86b28782cf49e4fe0ad8f0534be369612
Signed-off-by: Nicolas Trangez <ikke@nicolast.be>